+++ indico/modules/events/payment/notifications_test.py
@@ -8,9 +8,9 @@
 from pathlib import Path
 
 import pytest
-from flask import render_template
 
 from indico.modules.events.features.util import set_feature_enabled
+from indico.web.flask.templating import get_template_module
 
 
 pytest_plugins = 'indico.modules.events.registration.testing.fixtures'
@@ -21,7 +21,7 @@ def test_inconsistency_email_to_manager_plaintext(snapshot, dummy_event, dummy_r
     set_feature_enabled(dummy_event, 'registration', True)
     dummy_reg.base_price = 1337
     dummy_reg.currency = 'USD'
-    template = render_template('events/payment/emails/payment_inconsistency_email_to_manager.txt',
-                               event=dummy_event, registration=dummy_reg, amount=404, currency='EUR')
+    template = get_template_module('events/payment/emails/payment_inconsistency_email_to_manager.txt',
+                                   event=dummy_event, registration=dummy_reg, amount=404, currency='EUR')
     snapshot.snapshot_dir = Path(__file__).parent / 'templates/emails/tests'
-    snapshot.assert_match(template, 'payment_inconsistency_email_to_manager.txt')
+    snapshot.assert_match(template.get_body(), 'payment_inconsistency_email_to_manager.txt')