+++ indico/modules/rb/notifications/blockings.py
@@ -5,9 +5,8 @@
 # modify it under the terms of the MIT License; see the
 # LICENSE file for more details.
 
-from flask import render_template
-
 from indico.core.notifications import email_sender, make_email
+from indico.web.flask.templating import get_template_module
 
 
 @email_sender
@@ -16,10 +15,10 @@ def notify_request(owner, blocking, blocked_rooms):
 
     Expect only blockings for rooms owned by the specified owner.
     """
-    subject = 'Confirm room blockings'
-    body = render_template('rb/emails/blockings/awaiting_confirmation_email_to_manager.txt',
-                           owner=owner, blocking=blocking, blocked_rooms=blocked_rooms)
-    return make_email(owner.email, subject=subject, body=body)
+    with owner.force_user_locale():
+        tpl = get_template_module('rb/emails/blockings/awaiting_confirmation_email_to_manager.txt',
+                                  owner=owner, blocking=blocking, blocked_rooms=blocked_rooms)
+        return make_email(owner.email, template=tpl)
 
 
 @email_sender
@@ -28,9 +27,8 @@ def notify_request_response(blocked_room):
     Notify blocking creator about approval/rejection of his
     blocking request for a room.
     """
-    to = blocked_room.blocking.created_by_user.email
-    verb = blocked_room.State(blocked_room.state).title.upper()
-    subject = f'Room blocking {verb}'
-    body = render_template('rb/emails/blockings/state_email_to_user.txt',
-                           blocking=blocked_room.blocking, blocked_room=blocked_room, verb=verb)
-    return make_email(to, subject=subject, body=body)
+    user = blocked_room.blocking.created_by_user
+    with user.force_user_locale():
+        tpl = get_template_module('rb/emails/blockings/state_email_to_user.txt',
+                                  blocking=blocked_room.blocking, blocked_room=blocked_room)
+        return make_email(user.email, template=tpl)